requirements.txt
cryptography
qrcode
Pillow
pycryptodome